Update read me to add MK3 slicer bed model
Hi Jeff, I have posted a mk3 bed model to Thingiverse. Attributions listed in the TV page. If you like it, please feel free to use it. I am an ME rather than a coder, so my understanding of github is limited. I hope this was the correct way to suggest this to you :)